fix # of elements in wp list not matching pagination
Because of a LEFT JOIN, work package records can be present twice in the result returned by the SQL to fetch the work packages. As there is a chain of queries, having duplicates in the first query, which is used to fetch the ids for later querying will then result in those duplicates to be removed again when the work packages are actually filtered for (by the ids returned in the first query). But because of the duplicates begin removed, the number of elements does not match the number of specified elements, e.g. a only 35 elements returned for a pages size of 50.

Applying DISTINCT to the query for the ids solves the problem but requires to have all columns used for the ORDER BY present in the SELECT projection. This is done via a hack.